requirements.txt
asgiref==3.6.0
astroid==2.15.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dill==0.3.6
Django==4.2
djangorestframework==3.14.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.12.0
lazy-object-proxy==1.9.0
libretranslatepy==2.1.1
lxml==4.9.2
mccabe==0.7.0
multiprocess==0.70.14
pathos==0.3.0
platformdirs==3.2.0
pox==0.3.2
ppft==1.7.6.6
pycparser==2.21
PyExecJS==1.5.1
pylint==2.17.2
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.65.0
translate==3.6.1
translators==5.7.0
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
